#46 refactored service card into own component
